+++ crates/gpui/src/elements/mod.rs
@@ -5,12 +5,11 @@ mod deferred;
 mod div;
 mod image_cache;
 mod img;
+mod input;
 mod list;
 mod surface;
 mod svg;
 mod text;
-mod text_area;
-mod text_input;
 mod uniform_list;
 
 pub use anchored::*;
@@ -20,10 +19,9 @@ pub use deferred::*;
 pub use div::*;
 pub use image_cache::*;
 pub use img::*;
+pub use input::*;
 pub use list::*;
 pub use surface::*;
 pub use svg::*;
 pub use text::*;
-pub use text_area::*;
-pub use text_input::*;
 pub use uniform_list::*;
